[#4] Move basic parts of visualizer into module
