+++ source/application.ts
@@ -3,6 +3,7 @@ import './utilities/event-listener'
 import getUserMedia from './utilities/get-user-media'
 import log from './utilities/logger'
 import Poller from './utilities/poller'
+import renderBackground from './visualizers/default'
 
 const constraints = {
   audio: true,
@@ -25,44 +26,6 @@ const handleSuccess = () => {
     ))
     .then(() => setTimeout(pollAt240Seconds.stop, 240 * 1000 * 10))
 }
-
-const getCurrentSecond = () => {
-  const currentTime = new Date()
-  const currentSecond = currentTime.getHours() * 3600 +
-                        currentTime.getMinutes() * 60 +
-                        currentTime.getSeconds() +
-                        1 // Prevent a 0 value to allow for a simpler
-                          // representation of totalDegrees and totalSeconds.
-
-  log('info', `⏳ currentSecond: ${currentSecond}`)
-
-  return Promise.resolve(currentSecond)
-}
-
-const getCurrentHue = (currentSecond: number) => {
-  const totalDegrees = 360
-  const hueOffset = 240 // Start the day (midnight/0) at blue.
-  const totalSeconds = 86400 // Total number of seconds in a 24-hour period.
-
-  const hueStep = totalDegrees / totalSeconds
-  const currentStep = hueStep * currentSecond
-  const currentHue = (hueOffset + currentStep) % totalDegrees
-
-  log('info', `🎨 currentHue: ${currentHue}`)
-
-  return Promise.resolve(currentHue)
-}
-
-const updateBackgroundColor = (currentHue: number) => {
-  document.body.style.backgroundColor = `hsl(${currentHue}, 50%, 25%)`
-
-  log('info', `🖍 Background value: ${document.body.style.backgroundColor}`)
-}
-
-const renderBackground = () =>
-  getCurrentSecond()
-    .then(getCurrentHue)
-    .then(updateBackgroundColor)
 
 const logContactInfo = () =>
   log('info', `
